prototype-projects/backend/requirements.txt
asyncio==3.4.3
asyncpg==0.30.0
bcrypt==4.1.3
fastapi==0.111.0
fastapi-cli==0.0.4
pyjwt==2.10.1
ollama==0.2.1
pandas==2.2.3
#psycopg2==2.9.10
psycopg2-binary==2.9.9
#psycopg[binary]==3.2.6
pydantic==2.7.1
pydantic_core==2.18.2
requests==2.32.2
shapely==2.0.5
shellingham==1.5.4
#signxml==3.2.1
simple-websocket==1.0.0
uvicorn==0.22.0
uvloop==0.19.0
gridstatus==0.30.1
typesense==1.0.3